function deleteData

+++ src/services/bonsaisServe.js
@@ -16,5 +16,8 @@ export const getData = async () => {
 
   //DELETE
   export const deleteData = async (id) =>{
+    if(confirm("¿Estás seguro que quieres eliminar este bonsai?") === true)
     await fetch(`http://localhost:3000/bonsais/${id}`,{method:"delete"})
+    await getData();
+    location.reload();
   }
